api/requirements.txt
fastapi
uvicorn
pandas
numpy
tensorflow
keras
python-multipart
nest-asyncio
scikit-learn
requests
Pillow